Desacoplando testes dos modelos e observers, adicionando no teste da lib StatusVisNotification e EnrrolmentVisNotification. Closes #799
