+++ app/models/enrollment_observer.rb
@@ -1,5 +1,5 @@
 class EnrollmentObserver < ActiveRecord::Observer
-  include VisClient
+  include EnrollmentVisNotification
 
   def before_update(enrollment)
     old_enroll = Enrollment.find(enrollment.id)
@@ -9,27 +9,4 @@ class EnrollmentObserver < ActiveRecord::Observer
       notify_vis(enrollment, "remove_subject_finalized")
     end
   end
-
-  protected
-
-  def notify_vis(enrollment, type)
-    params = {
-      :user_id => enrollment.user_id,
-      :lecture_id => nil,
-      :subject_id => enrollment.subject_id,
-      :space_id => enrollment.subject.space.id,
-      :course_id => enrollment.subject.space.course.id,
-      :type => type,
-      :status_id => nil,
-      :statusable_id => nil,
-      :statusable_type => nil,
-      :in_response_to_id => nil,
-      :in_response_to_type => nil,
-      :created_at => enrollment.created_at,
-      :updated_at => enrollment.updated_at
-    }
-
-    self.send_async_info(params, Redu::Application.config.vis_client[:url])
-  end
-
 end
